small config changes

+++ llm_summarize/SFT/SFT.py
@@ -1,4 +1,3 @@
-import torch
 from transformers import AutoModelForCausalLM, EarlyStoppingCallback
 from llm_summarize.SFT.custom_sfttrainer import CustomSFTTrainer
 import wandb
@@ -18,7 +17,7 @@ def run_SFT(train_dataset: Dataset, eval_dataset: Dataset, config: MainConfig) -
     model = AutoModelForCausalLM.from_pretrained(
         config.model.model_name,
         quantization_config=quantization_config,
-        torch_dtype=config.model.dtype,
+        dtype = config.model.dtype,
         attn_implementation=config.model.attn_implementation,
         device_map=config.model.device_map
     )
